+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Engine.java
@@ -997,10 +997,6 @@ public final class Engine implements AutoCloseable {
             return new Object();
         }
 
-        @Override
-        public SecurityException throwSecurityException(String message) {
-            throw new SecurityException(message);
-        }
     }
 
 }