+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleLanguage.java
@@ -2134,7 +2134,7 @@ public abstract class TruffleLanguage<C> {
          * environment.
          *
          * @see #newProcessBuilder(java.lang.String...)
-         * @since 20.0.0 beta 1
+         * @since 20.0.0 beta 2
          */
         public boolean isCreateProcessAllowed() {
             return AccessAPI.engineAccess().isCreateProcessAllowed(vmObject);
@@ -2145,7 +2145,7 @@ public abstract class TruffleLanguage<C> {
          *
          * @param command the executable and its arguments
          * @throws SecurityException when process creation is not allowed
-         * @since 20.0.0 beta 1
+         * @since 20.0.0 beta 2
          */
         @TruffleBoundary
         public TruffleProcessBuilder newProcessBuilder(String... command) {
@@ -2163,7 +2163,7 @@ public abstract class TruffleLanguage<C> {
          * {@link EnvironmentAccess#NONE} an empty map is returned.
          *
          * @return the process environment as a map of variable names to values
-         * @since 20.0.0 beta 1
+         * @since 20.0.0 beta 2
          */
         @TruffleBoundary
         public Map<String, String> getEnvironment() {