+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/io/package-info.java
@@ -49,6 +49,6 @@
  * This package contains I/O related classes used by {@link com.oracle.truffle.api.TruffleLanguage}
  * implementations.
  *
- * @since 20.0.0 beta 1
+ * @since 20.0.0 beta 2
  */
 package com.oracle.truffle.api.io;